handle already followed person and not exsits person

+++ src/main/java/com/codecool/poster/model/Media.java
@@ -13,9 +13,10 @@ import javax.persistence.*;
 @Builder
 @Entity
 public class Media {
+
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private int id;
+    private long id;
 
     @ManyToOne
     @JoinColumn(referencedColumnName = "id")
